Change from int to long for values that represent sizes in bit
This fixes parsing of traces with big packets, whose size in bits
overflow an 32-bit int.

Reviewed-on: https://git.eclipse.org/r/17422
Tested-by: Hudson CI
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>

+++ org.eclipse.linuxtools.ctf.core.tests/src/org/eclipse/linuxtools/ctf/core/tests/io/BitBufferTest.java
@@ -143,7 +143,7 @@ public class BitBufferTest {
      */
     @Test
     public void testGetPosition() {
-        int result = fixture.position();
+        long result = fixture.position();
 
         assertEquals(1, result);
     }
