+++ org.eclipse.linuxtools.ctf.core/src/org/eclipse/linuxtools/ctf/core/event/io/BitBuffer.java
@@ -46,7 +46,7 @@ public final class BitBuffer {
     // ------------------------------------------------------------------------
 
     private ByteBuffer buf;
-    private int pos;
+    private long pos;
     private ByteOrder byteOrder;
 
     // ------------------------------------------------------------------------
@@ -128,9 +128,9 @@ public final class BitBuffer {
             case BitBuffer.BIT_CHAR:
                 // Byte
                 if (signed) {
-                    val = this.buf.get(this.pos / 8);
+                    val = this.buf.get((int) (this.pos / 8));
                 } else {
-                    val = (this.buf.get(this.pos / 8)) & 0xff;
+                    val = (this.buf.get((int) (this.pos / 8))) & 0xff;
                 }
                 gotIt = true;
                 break;
@@ -138,9 +138,9 @@ public final class BitBuffer {
             case BitBuffer.BIT_SHORT:
                 // Word
                 if (signed) {
-                    val = this.buf.getShort(this.pos / 8);
+                    val = this.buf.getShort((int) (this.pos / 8));
                 } else {
-                    short a = this.buf.getShort(this.pos / 8);
+                    short a = this.buf.getShort((int) (this.pos / 8));
                     val = a & 0xffff;
                 }
                 gotIt = true;
@@ -148,7 +148,7 @@ public final class BitBuffer {
 
             case BitBuffer.BIT_INT:
                 // Double word
-                val = this.buf.getInt(this.pos / 8);
+                val = this.buf.getInt((int) (this.pos / 8));
                 gotIt = true;
                 break;
 
@@ -169,11 +169,11 @@ public final class BitBuffer {
         return val;
     }
 
-    private int getIntBE(int index, int length, boolean signed) {
+    private int getIntBE(long index, int length, boolean signed) {
         assert ((length > 0) && (length <= BIT_INT));
-        int end = index + length;
-        int startByte = index / BIT_CHAR;
-        int endByte = (end + (BIT_CHAR - 1)) / BIT_CHAR;
+        long end = index + length;
+        int startByte = (int) (index / BIT_CHAR);
+        int endByte = (int) ((end + (BIT_CHAR - 1)) / BIT_CHAR);
         int currByte, lshift, cshift, mask, cmask, cache;
         int value = 0;
 
@@ -193,7 +193,7 @@ public final class BitBuffer {
             value |= cmask;
             return value;
         }
-        cshift = index % BIT_CHAR;
+        cshift = (int) (index % BIT_CHAR);
         if (cshift > 0) {
             mask = ~((~0) << (BIT_CHAR - cshift));
             cmask = cache & mask;
@@ -206,7 +206,7 @@ public final class BitBuffer {
             value <<= BIT_CHAR;
             value |= this.buf.get(currByte) & 0xFF;
         }
-        lshift = end % BIT_CHAR;
+        lshift = (int) (end % BIT_CHAR);
         if (lshift > 0) {
             mask = ~((~0) << lshift);
             cmask = this.buf.get(currByte) & 0xFF;
@@ -221,17 +221,17 @@ public final class BitBuffer {
         return value;
     }
 
-    private int getIntLE(int index, int length, boolean signed) {
+    private int getIntLE(long index, int length, boolean signed) {
         assert ((length > 0) && (length <= BIT_INT));
-        int end = index + length;
-        int startByte = index / BIT_CHAR;
-        int endByte = (end + (BIT_CHAR - 1)) / BIT_CHAR;
+        long end = index + length;
+        int startByte = (int) (index / BIT_CHAR);
+        int endByte = (int) ((end + (BIT_CHAR - 1)) / BIT_CHAR);
         int currByte, lshift, cshift, mask, cmask, cache, mod;
         int value = 0;
 
         currByte = endByte - 1;
         cache = buf.get(currByte) & 0xFF;
-        mod = end % BIT_CHAR;
+        mod = (int) (end % BIT_CHAR);
         lshift = (mod > 0) ? mod : BIT_CHAR;
         boolean isNeg = (cache & (1 << (lshift - 1))) != 0;
         if (signed && isNeg) {
@@ -247,7 +247,7 @@ public final class BitBuffer {
             value |= cmask;
             return value;
         }
-        cshift = end % BIT_CHAR;
+        cshift = (int) (end % BIT_CHAR);
         if (cshift > 0) {
             mask = ~((~0) << cshift);
             cmask = cache & mask;
@@ -259,7 +259,7 @@ public final class BitBuffer {
             value <<= BIT_CHAR;
             value |= buf.get(currByte) & 0xFF;
         }
-        lshift = index % BIT_CHAR;
+        lshift = (int) (index % BIT_CHAR);
         if (lshift > 0) {
             mask = ~((~0) << (BIT_CHAR - lshift));
             cmask = buf.get(currByte) & 0xFF;
@@ -307,7 +307,7 @@ public final class BitBuffer {
      *            The value to write
      */
     public void putInt(int length, int value) {
-        final int curPos = this.pos;
+        final long curPos = this.pos;
 
         if (!canRead(length)) {
             throw new BufferOverflowException();
@@ -323,11 +323,11 @@ public final class BitBuffer {
         this.pos += length;
     }
 
-    private void putIntBE(int index, int length, int value) {
+    private void putIntBE(long index, int length, int value) {
         assert ((length > 0) && (length <= BIT_INT));
-        int end = index + length;
-        int startByte = index / BIT_CHAR;
-        int endByte = (end + (BIT_CHAR - 1)) / BIT_CHAR;
+        long end = index + length;
+        int startByte = (int) (index / BIT_CHAR);
+        int endByte = (int) ((end + (BIT_CHAR - 1)) / BIT_CHAR);
         int currByte, lshift, cshift, mask, cmask;
         int correctedValue = value;
 
@@ -342,7 +342,7 @@ public final class BitBuffer {
 
         /* sub byte */
         if (startByte == (endByte - 1)) {
-            lshift = (BIT_CHAR - (end % BIT_CHAR)) % BIT_CHAR;
+            lshift = (int) ((BIT_CHAR - (end % BIT_CHAR)) % BIT_CHAR);
             mask = ~((~0) << lshift);
             if ((index % BIT_CHAR) > 0) {
                 mask |= (~(0)) << (BIT_CHAR - (index % BIT_CHAR));
@@ -360,7 +360,7 @@ public final class BitBuffer {
 
         /* head byte contains MSB */
         currByte = endByte - 1;
-        cshift = end % BIT_CHAR;
+        cshift = (int) (end % BIT_CHAR);
         if (cshift > 0) {
             lshift = BIT_CHAR - cshift;
             mask = ~((~0) << lshift);
@@ -388,11 +388,11 @@ public final class BitBuffer {
         }
     }
 
-    private void putIntLE(int index, int length, int value) {
+    private void putIntLE(long index, int length, int value) {
         assert ((length > 0) && (length <= BIT_INT));
-        int end = index + length;
-        int startByte = index / BIT_CHAR;
-        int endByte = (end + (BIT_CHAR - 1)) / BIT_CHAR;
+        long end = index + length;
+        int startByte = (int) (index / BIT_CHAR);
+        int endByte = (int) ((end + (BIT_CHAR - 1)) / BIT_CHAR);
         int currByte, lshift, cshift, mask, cmask;
         int correctedValue = value;
 
@@ -407,7 +407,7 @@ public final class BitBuffer {
 
         /* sub byte */
         if (startByte == (endByte - 1)) {
-            lshift = index % BIT_CHAR;
+            lshift = (int) (index % BIT_CHAR);
             mask = ~((~0) << lshift);
             if ((end % BIT_CHAR) > 0) {
                 mask |= (~(0)) << (end % BIT_CHAR);
@@ -425,7 +425,7 @@ public final class BitBuffer {
 
         /* head byte */
         currByte = startByte;
-        cshift = index % BIT_CHAR;
+        cshift = (int) (index % BIT_CHAR);
         if (cshift > 0) {
             mask = ~((~0) << cshift);
             cmask = correctedValue << cshift;
@@ -468,7 +468,7 @@ public final class BitBuffer {
             return false;
         }
 
-        if ((this.pos + length) > (this.buf.capacity() * BIT_CHAR)) {
+        if ((this.pos + length) > (((long) this.buf.capacity()) * BIT_CHAR)) {
             return false;
         }
         return true;
@@ -502,7 +502,7 @@ public final class BitBuffer {
      * @param newPosition
      *            The new position of the buffer.
      */
-    public void position(int newPosition) {
+    public void position(long newPosition) {
         this.pos = newPosition;
     }
 
@@ -512,7 +512,7 @@ public final class BitBuffer {
      *
      * @return order The position of the buffer.
      */
-    public int position() {
+    public long position() {
         return this.pos;
     }
