+++ org.eclipse.linuxtools.ctf.core/src/org/eclipse/linuxtools/ctf/core/event/types/EnumDefinition.java
@@ -104,7 +104,7 @@ public class EnumDefinition extends SimpleDatatypeDefinition {
     @Override
     public void read(BitBuffer input) {
         int align = (int) declaration.getAlignment();
-        int pos = input.position() + ((align-(input.position() % align))%align);
+        long pos = input.position() + ((align - (input.position() % align)) % align);
         input.position(pos);
         integerValue.read(input);
         long val = integerValue.getValue();
